Fix storybook
