+++ package.json
@@ -64,6 +64,7 @@
     "react": "^16.0.0",
     "react-aspect-ratio": "^1.0.3",
     "react-dom": "^16.0.0",
+    "react-storybook-addon-chapters": "^3.1.3",
     "react-test-renderer": "^16.0.0",
     "storybook-addon-jsx": "^7.0.0"
   },
